% Generated by roxygen2 (4.0.1): do not edit by hand
\name{beePlot}
\alias{beePlot}
\title{Create a beeswarm Plot}
\usage{
beePlot(id, category, praemie, rank, data, color, range = c(0, max),
  conf = T, median = T, mean = F, label = T, bee_plot = T, cex = 6,
  cex.axis = 1.8, cex.lab = 1.4, cex.med = 2.2, cex.minmax = 1.3,
  lwd.med = 3, las = 1, method = "center", corral = "random")
}
\arguments{
\item{id}{}

\item{category}{}

\item{praemie}{}

\item{rank}{}

\item{data}{}

\item{color}{}

\item{range}{}

\item{conf}{}

\item{median}{}

\item{mean}{}

\item{label}{}

\item{bee_plot}{}

\item{cex}{}

\item{cex.axis}{}

\item{cex.lab}{}

\item{cex.med}{}

\item{cex.minmax}{}

\item{lwd.med}{}

\item{method}{}

\item{corral}{}

\item{las}{}
}
\description{
\code{beePlot}
}

